Move `required_state` tests
